+++ b4n/core/app.rs
@@ -411,7 +411,7 @@ impl App {
     /// Changes application theme.
     fn process_theme_change(&mut self, theme: String) {
         if self.data.borrow().config.theme != theme {
-            let msg = format!("Theme changed to '{theme}'…");
+            let msg = format!("Theme changed to '{theme}'");
             self.data.borrow_mut().config.theme = theme;
             let _ = self.theme_watcher.change_file(self.data.borrow().config.theme_path());
             self.config_watcher.skip_next();