+++ b4n/ui/views/logs/view.rs
@@ -100,6 +100,7 @@ impl LogsView {
         self.command_palette =
             CommandPalette::new(Rc::clone(&self.app_data), actions, 65).with_highlighted_position(self.last_mouse_click.take());
         self.command_palette.show();
+        self.footer.hide_hint();
     }
 
     fn show_mouse_menu(&mut self, x: u16, y: u16) {
@@ -136,12 +137,12 @@ impl LogsView {
                     .is_ok()
             {
                 if self.logs.has_selection() {
-                    self.footer.show_info(" Selection copied to clipboard…", 3_000);
+                    self.footer.show_info("Selection copied to clipboard", 3_000);
                 } else {
-                    self.footer.show_info(" Container logs copied to clipboard…", 3_000);
+                    self.footer.show_info("Container logs copied to clipboard", 3_000);
                 }
             } else {
-                self.footer.show_error(" Unable to access clipboard functionality…", 5_000);
+                self.footer.show_error("Unable to access clipboard functionality", 5_000);
             }
         }
     }